move useWindowSize logic to a custom hook
