+++ src/hooks/useWindowSize.js
@@ -1,0 +1,20 @@
+/* eslint-disable */
+
+import React, { useEffect, useState } from 'react';
+
+// custom hook to watch for window size
+
+export default function useWindowSize() {
+  const [size, setSize] = useState([window.innerHeight, window.innerWidth]);
+  useEffect(() => {
+    const handleResize = () => {
+      setSize([window.innerHeight, window.innerWidth]);
+    };
+    window.addEventListener('resize', handleResize);
+    return () => {
+      window.removeEventListener('resize', handleResize);
+    };
+  }, []);
+  return size;
+};
+
